+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated [`@adguard/scriptlets`] to v1.9.96
+- Updated [`@adguard/scriptlets`] to v1.9.105
 
 
 ## [1.11.16] - 2023-11-21